Adress PR comments
