+++ tests/integrations/test_db_integrations_sql.py
@@ -4,7 +4,7 @@
 from utils import context, bug, missing_feature, irrelevant, scenarios, flaky
 from utils.tools import logger
 
-from .integrations_utils import BaseDbIntegrationsTestClass, get_span_from_agent, get_span_from_tracer
+from .sql_utils import BaseDbIntegrationsTestClass
 
 
 class _BaseDatadogDbIntegrationTestClass(BaseDbIntegrationsTestClass):
@@ -20,10 +20,10 @@ class _BaseDatadogDbIntegrationTestClass(BaseDbIntegrationsTestClass):
             logger.debug(f"Start validation for {self.db_service}/{db_operation}")
 
             logger.debug("Validating span generated by Datadog library")
-            yield db_operation, get_span_from_tracer(request)
+            yield db_operation, self.get_span_from_tracer(request)
 
             logger.debug("Validating transmitted by Datadog agent")
-            yield db_operation, get_span_from_agent(request)
+            yield db_operation, self.get_span_from_agent(request)
 
     # Tests methods
     @flaky(context.library >= "java@1.23.0", reason="DBMON-3088")
@@ -202,13 +202,13 @@ class _BaseDatadogDbIntegrationTestClass(BaseDbIntegrationsTestClass):
     def test_NOT_obfuscate_query(self):
         """ All queries come out without obfuscation from tracer library """
         for db_operation, request in self.get_requests():
-            span = get_span_from_tracer(request)
+            span = self.get_span_from_tracer(request)
             assert span["resource"].count("?") == 0, f"The query should not be obfuscated for operation {db_operation}"
 
     def test_sql_query(self):
         """ Usually the query """
         for db_operation, request in self.get_requests(excluded_operations=["procedure", "select_error"]):
-            span = get_span_from_agent(request)
+            span = self.get_span_from_agent(request)
             assert (
                 db_operation in span["meta"]["sql.query"].lower()
             ), f"sql.query span not found for operation {db_operation}"
@@ -217,7 +217,7 @@ class _BaseDatadogDbIntegrationTestClass(BaseDbIntegrationsTestClass):
         """ All queries come out obfuscated from agent """
         for db_operation, request in self.get_requests(excluded_operations=excluded_operations):
 
-            span = get_span_from_agent(request)
+            span = self.get_span_from_agent(request)
             # We launch all queries with two parameters (from weblog)
             # Insert and procedure:These operations also receive two parameters, but are obfuscated as only one.
             if db_operation in ["insert", "procedure"]:
@@ -341,7 +341,7 @@ class Test_MsSql(_BaseDatadogDbIntegrationTestClass):
         """ All queries come out obfuscated from agent """
         for db_operation, request in self.get_requests(excluded_operations=excluded_operations):
 
-            span = get_span_from_agent(request)
+            span = self.get_span_from_agent(request)
             # We launch all queries with two parameters (from weblog)
             if db_operation == "insert":
                 expected_obfuscation_count = 1
